Create MersennePrime.py
This is a Mersenne Prime number finder that I am still trying to implement multiprocessing in.
This is not multiprocessing because the algorithm requires a previous iteration of modular division that 
cannot be looked ahead on to be set up. Any attempts I have made to make it multiprocessing have only
made it slower. This program only calculates Mersenne Prime numbers: (2^n) -1